requirements.txt
Pillow>=6
numpy>=1.16
requests
split-image
opencv-python